requirements.txt
appdirs==1.4.3
autopep8==1.5.4
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==2.1.0
ecdsa==0.14.1
fastapi==0.63.0
fastapi-contrib==0.2.9
firebase-admin==4.5.1
flake8==3.8.4
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
h11==0.12.0
html5lib==1.0.1
httplib2==0.18.1
httptools==0.1.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
mccabe==0.6.1
motor==2.3.0
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pyflakes==2.2.0
pymongo==3.11.2
pyparsing==2.4.6
python-dotenv==0.15.0
python-jose==3.2.0
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
rsa==4.7
six==1.14.0
starlette==0.13.6
toml==0.10.2
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
webencodings==0.5.1
websockets==8.1
